+++ xradar/io/export/__init__.py
@@ -15,6 +15,7 @@ Data Export
 """
 
 from .cfradial2 import *  # noqa
+from .cfradial1 import *  # noqa
 from .odim import *  # noqa
 
 __all__ = [s for s in dir() if not s.startswith("_")]