+++ xradar/io/export/odim.py
@@ -59,7 +59,7 @@ def _write_odim(source, destination):
             destination.attrs[key] = value
 
 
-def _write_odim_dataspace(source, destination):
+def _write_odim_dataspace(source, destination, compression, compression_opts):
     """Write ODIM_H5 Dataspaces.
 
     Parameters
@@ -68,6 +68,10 @@ def _write_odim_dataspace(source, destination):
         Moments to write
     destination : handle
         h5py-group handle
+    compression : str
+        Compression filter name
+    compression_opts : compression strategy
+        options as needed by above filter
     """
     # todo: check bottom-up/top-down rhi
     dim0 = "elevation" if source.sweep_mode == "rhi" else "azimuth"
@@ -105,17 +109,15 @@ def _write_odim_dataspace(source, destination):
         val = value.sortby(dim0).values
         fillval = _fillvalue * scale_factor
         fillval += add_offset
-        val[np.isnan(val)] = fillval
         val = (val - add_offset) / scale_factor
+        val[np.isnan(val)] = fillval
         if np.issubdtype(dtype, np.integer):
             val = np.rint(val).astype(dtype)
-        # todo: compression is chosen totally arbitrary here
-        #  maybe parameterizing it?
         ds = h5_data.create_dataset(
             "data",
             data=val,
-            compression="gzip",
-            compression_opts=6,
+            compression=compression,
+            compression_opts=compression_opts,
             fillvalue=_fillvalue,
             dtype=dtype,
         )
@@ -132,7 +134,14 @@ def _write_odim_dataspace(source, destination):
             ds.attrs.create("IMAGE_VERSION", version, dtype=H5T_C_S1_VER)
 
 
-def to_odim(dtree, filename):
+def to_odim(
+    dtree,
+    filename,
+    source=None,
+    optional_how=False,
+    compression="gzip",
+    compression_opts=6,
+):
     """Save DataTree to ODIM_H5/V2_2 compliant file.
 
     Parameters
@@ -140,7 +149,27 @@ def to_odim(dtree, filename):
     dtree : datatree.DataTree
     filename : str
         output filename
+
+    Keyword Arguments
+    -----------------
+    source : str
+        mandatory radar identifier (see ODIM documentation)
+    optional_how : boolean
+        True to include optional how attributes, defaults to False
+    compression : str
+        Compression filter name, defaults to "gzip".
+    compression_opts : compression strategy
+        options as needed by above filter, defaults to 6
     """
+    has_identifier = False
+    if source is not None:
+        has_identifier = any(key in source for key in ["NOD", "WMO", "RAD"])
+    if not has_identifier:
+        raise ValueError(
+            "Please provide the source parameter with at least one"
+            "of the mandatory radar identifier (NOD, RAD, WMO)"
+        )
+
     root = dtree["/"]
 
     h5 = h5py.File(filename, "w")
@@ -168,7 +197,8 @@ def to_odim(dtree, filename):
     what["version"] = "H5rad 2.2"
     what["date"] = str(root["time_coverage_start"].values)[:10].replace("-", "")
     what["time"] = str(root["time_coverage_end"].values)[11:19].replace(":", "")
-    what["source"] = root.attrs["instrument_name"]
+
+    what["source"] = source
 
     h5_what = h5.create_group("what")
     _write_odim(what, h5_what)
@@ -244,19 +274,23 @@ def to_odim(dtree, filename):
 
         # ODIM_H5 datasetN numbers are 1-based
         sweep_number = ds.sweep_number + 1
-        ds_how = {
-            "scan_index": sweep_number,
-            "scan_count": len(grps),
-            "startazT": tout - difft,
-            "stopazT": tout + difft,
-            "startazA": azout - diffa,
-            "stopazA": azout + diffa,
-            "startelA": elout - diffe,
-            "stopelA": elout + diffe,
-        }
+
+        ds_how = {}
+        if optional_how:
+            optional = {
+                "scan_index": sweep_number,
+                "scan_count": len(grps),
+                "startazT": tout - difft,
+                "stopazT": tout + difft,
+                "startazA": azout - diffa,
+                "stopazA": azout + diffa,
+                "startelA": elout - diffe,
+                "stopelA": elout + diffe,
+            }
+            ds_how.update(optional)
         _write_odim(ds_how, h5_ds_how)
 
         # write moments
-        _write_odim_dataspace(ds, h5_dataset)
+        _write_odim_dataspace(ds, h5_dataset, compression, compression_opts)
 
     h5.close()